Create some stub modules to make our python on NAO happy.
Since we are using a cross-built Python 3.10 on NAO, not all the
packages have been installed via Pip, this results in stops when
installing our gimmick_model package. We have all these installed,
just not through pip. So, this is the minimum stubs so that pip
doesn't complain when installing gimmick_model.

One *should not* use these if one can easily get the real modules.
This only places an entry in pip's database. We assume the items
are installed through other ways.