Try making a version that can be run remotely.
This means that we can run the program on older robots, like NAOv5,
and maybe Pepper if we squint enough.

Since I was here, I also took the chance to restructure the code. This
probably needs to do a separate case for the "client" case, but I have
ideas for that too.